requirements.txt
Flask
requests
python-dateutil
PyPDF2
instaloader
gTTS
qrcode
googletrans==4.0.0-rc1
pillow
